[REF] web, web_editor: remove service_mixins.js
To discourage the use of that mixin outside of its remaining usecase and
as a new step towards the future refactoring of public widgets, now move
that mixin directly in the public widgets file.

Related to task-3930204
