+++ addons/web/static/src/legacy/js/public/public_widget.js
@@ -2,10 +2,11 @@
  * Provides a way to start JS code for public contents.
  */
 
+import { Component } from "@odoo/owl";
 import dom from '@web/legacy/js/core/dom';
 import Class from "@web/legacy/js/core/class";
-import ServicesMixin from "@web/legacy/js/core/service_mixins";
 import { loadBundle, loadCSS, loadJS } from '@web/core/assets';
+import { SERVICES_METADATA } from "@web/core/utils/hooks";
 import { renderToElement } from "@web/core/utils/render";
 import { makeAsyncHandler, makeButtonHandler } from "@web/legacy/js/public/minimal_dom";
 
@@ -313,6 +314,65 @@ const EventDispatcherMixin = Object.assign({}, ParentedMixin, {
         ParentedMixin.destroy.call(this);
     },
 });
+
+function protectMethod(widget, fn) {
+    return function (...args) {
+        return new Promise((resolve, reject) => {
+            Promise.resolve(fn.call(this, ...args))
+                .then((result) => {
+                    if (!widget.isDestroyed()) {
+                        resolve(result);
+                    }
+                })
+                .catch((reason) => {
+                    if (!widget.isDestroyed()) {
+                        reject(reason);
+                    }
+                });
+        });
+    };
+}
+
+const ServicesMixin = {
+    bindService: function (serviceName) {
+        const { services } = Component.env;
+        const service = services[serviceName];
+        if (!service) {
+            throw new Error(`Service ${serviceName} is not available`);
+        }
+        if (serviceName in SERVICES_METADATA) {
+            if (service instanceof Function) {
+                return protectMethod(this, service);
+            } else {
+                const methods = SERVICES_METADATA[serviceName];
+                const result = Object.create(service);
+                for (const method of methods) {
+                    result[method] = protectMethod(this, service[method]);
+                }
+                return result;
+            }
+        }
+        return service;
+    },
+    /**
+     * @param  {string} service
+     * @param  {string} method
+     * @return {any} result of the service called
+     */
+    call: function (service, method) {
+        var args = Array.prototype.slice.call(arguments, 2);
+        var result;
+        this.trigger_up('call_service', {
+            service: service,
+            method: method,
+            args: args,
+            callback: function (r) {
+                result = r;
+            },
+        });
+        return result;
+    },
+};
 
 /**
  * Base class for all visual components. Provides a lot of functions helpful
@@ -918,4 +978,5 @@ export default {
 
     ParentedMixin: ParentedMixin,
     EventDispatcherMixin: EventDispatcherMixin,
+    ServicesMixin: ServicesMixin,
 };
